code/requirements.txt
psycopg2-binary
matplotlib
tqdm
wget
xlrd
openpyxl
pandas